category route and controller test 1

+++ database/migrations/2024_01_08_095756_create_cards_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->foreignId('rarity_id');
             $table->foreignId('user_id');
             $table->string('image')->nullable();
-            $table->integer('active');
+            $table->tinyInteger('active');
             $table->timestamps();
         });
     }
